refactor(run): the epilogue earns its seam — the run's last words in one module
Pure move, byte-identical behavior: print_resume_summary ·
print_flow_epilogue · outputs_note · json_type_name ·
outputs_json_line · emit_diagnostic · emit_error_envelope ·
paused_envelope_line · error_envelope_line · first_nika_code ·
envelope_message · run_failure_envelope leave run/mod.rs for the new
run/epilogue.rs (pub(super) · their tests ride along). trace_sink
stays — it is run-START composition, not epilogue. The fused doc
comment above trace_sink split back to its two owners. run/mod.rs
1487 → 1196 lines; the display features of #319/#321 now have one
seam to land in, never the driver.
